Some updates per PR feedback.
